Added lots of things, changed some powerup rarities, added some finishing touches to the PVP Battle, PVP now also spawns waves of enemies if it takes too long

+++ Assets/ScriptableObjects/PlayerClasses/PlayerClass.cs
@@ -40,7 +40,7 @@ public class PlayerClass : ScriptableObject
     public float bulletSizeDelta = 0;
     public float bulletSpeedDelta = 0;
     public bool splitOnHit = false;
-    public float splitAmountDelta = 0;
+    public int splitAmountDelta = 0;
     public float splitRangeDelta = 0;
     public float splitBulletSizeDelta = 0;
     public float splitBulletSpeedDelta = 0;
